third_party/WebKit/LayoutTests/http/tests/sendbeacon/beacon-allowance.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Test navigator.sendBeacon() restrictions on (accumulated) transmission length.");

var payload = new Uint32Array(1000);
var result = true;
function testAllowance() {
    for (var i = 0 ; result && i < 100; i++)
        result = navigator.sendBeacon("resources/blank.txt", payload);
    shouldBeFalse('result');
    finishJSTest();
}

if (window.testRunner) {
    testRunner.dumpAsText();
}
testAllowance();
</script>
</head>
</html>
