third_party/blink/web_tests/http/tests/sendbeacon/beacon-allowance-expected.txt

Test navigator.sendBeacon() restrictions on (accumulated) transmission length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is false
PASS successfullyParsed is true

TEST COMPLETE
